auction_system/requirements.txt
django==3.1.12
djongo==1.3.7
sqlparse==0.2.4
djangorestframework==3.12.4
channels==3.0.5
channels-redis==3.4.1
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.11.0
daphne
pytz